Foundation: avoid temporary file leaks on Windows
Correct the implementation of `_NSCleanupTemporaryFile` to use
`MoveFileExW` with `MOVEFILE_COPY_ALLOWED` to prefer to use file renames
rather than copy and delete.  This should improve performance in the
general single volume case and still allow the correct behaviour in the
multi-volume scenarios using `CopyFile` and `DeleteFile`.

Thanks to @darinf for identifying the test case with 65535 temporary
files!

+++ Sources/Foundation/FileManager+Win32.swift
@@ -620,7 +620,7 @@ extension FileManager {
                 // case, we need to do a recursive copy & remove.
                 if PathIsSameRootW(wszSource, wszDestination) ||
                         faSourceAttributes.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY == 0 {
-                    if !MoveFileExW(wszSource, wszDestination, DWORD(MOVEFILE_COPY_ALLOWED | MOVEFILE_WRITE_THROUGH)) {
+                    if !MoveFileExW(wszSource, wszDestination, MOVEFILE_COPY_ALLOWED | MOVEFILE_WRITE_THROUGH) {
                         throw _NSErrorWithWindowsError(GetLastError(), reading: false, paths: [srcPath, dstPath])
                     }
                 } else {
